Work on Debugger (#255)
* Work on Debugger

- plug Debugger with WireProtocol in plain C
- fix #226
- completly replaced WireProtocol C++ with C version
- replace private_malloc with platform_malloc
- rework Monitor_FlashSectorMap to report full details on BlockStorage
- fixes #228
- further improvements in BlockStorage
- add stubs for COM
- add interface from BlockStorage interface to STM32 Flash driver
- rework nanoCLR VS Solution to make it compatible with latest changes
- revert several types to well know types
- add weak declarations to COM stubs in Win32
- correct linker scripts to have addresses aligned with sector boundaries
- add block storage definitions for all reference boards
- delete CMake file for NF_WireProtocol (is already in NF_Debugger)

+++ src/HAL/HAL_Time.cpp
@@ -22,19 +22,19 @@ const int CummulativeDaysForMonth[13] = {0, 31, 59, 90, 120, 151, 181, 212, 243,
 #define MINUTES_TO_HOUR             60
 #define HOURS_TO_DAY                24
 
-signed __int64  HAL_Time_CurrentTime() { 
+signed long long  HAL_Time_CurrentTime() { 
     return HAL_Time_SysTicksToTime( HAL_Time_CurrentSysTicks() ); 
 };
 
-signed __int64 HAL_Time_FromSystemTime(const SYSTEMTIME* systemTime)
+signed long long HAL_Time_FromSystemTime(const SYSTEMTIME* systemTime)
 {
-    signed __int64 r = YEARS_TO_DAYS(systemTime->wYear) + MONTH_TO_DAYS(systemTime->wYear, systemTime->wMonth) + systemTime->wDay - 1;
+    signed long long r = YEARS_TO_DAYS(systemTime->wYear) + MONTH_TO_DAYS(systemTime->wYear, systemTime->wMonth) + systemTime->wDay - 1;
     r = (((( (r * HOURS_TO_DAY) + systemTime->wHour) * MINUTES_TO_HOUR + systemTime->wMinute) * SECONDS_TO_MINUTES + systemTime->wSecond ) * MILLISECONDS_TO_SECONDS + systemTime->wMilliseconds) * TIMEUNIT_TO_MILLISECONDS;
 
     return r;    
 }
 
-bool HAL_Time_ToSystemTime(signed __int64 time, SYSTEMTIME* systemTime)
+bool HAL_Time_ToSystemTime(signed long long time, SYSTEMTIME* systemTime)
 {
     int ytd = 0;
     int mtd = 0;
